Move ingredient and priority handling out of GristCostRecipe to new abstract class
