+++ src/main/java/com/mraof/minestuck/alchemy/recipe/GristCostRecipe.java
@@ -1,26 +1,22 @@
 package com.mraof.minestuck.alchemy.recipe;
 
-import com.google.gson.JsonObject;
-import com.mraof.minestuck.api.alchemy.GristSet;
-import com.mraof.minestuck.api.alchemy.MutableGristSet;
 import com.mraof.minestuck.alchemy.recipe.generator.GeneratedCostProvider;
-import com.mraof.minestuck.alchemy.recipe.generator.GenerationContext;
-import com.mraof.minestuck.alchemy.recipe.generator.GristCostResult;
+import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.api.alchemy.GristType;
-import com.mraof.minestuck.api.alchemy.GristTypes;
+import com.mraof.minestuck.api.alchemy.MutableGristSet;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
 import com.mraof.minestuck.jei.JeiGristCost;
 import net.minecraft.MethodsReturnNonnullByDefault;
-import net.minecraft.core.NonNullList;
 import net.minecraft.core.RegistryAccess;
-import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.GsonHelper;
 import net.minecraft.world.Container;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.*;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeManager;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.Level;
 
 import javax.annotation.Nullable;
@@ -52,21 +48,10 @@ public abstract class GristCostRecipe implements Recipe<Container>
 	}
 	
 	public final ResourceLocation id;
-	protected final Ingredient ingredient;
-	@Nullable
-	protected final Integer priority;
 	
-	public GristCostRecipe(ResourceLocation id, Ingredient ingredient, @Nullable Integer priority)
+	public GristCostRecipe(ResourceLocation id)
 	{
 		this.id = id;
-		this.ingredient = ingredient;
-		this.priority = priority;
-	}
-	
-	@Override
-	public boolean matches(Container inv, Level level)
-	{
-		return ingredient.test(inv.getItem(0));
 	}
 	
 	@Override
@@ -94,12 +79,6 @@ public abstract class GristCostRecipe implements Recipe<Container>
 	}
 	
 	@Override
-	public NonNullList<Ingredient> getIngredients()
-	{
-		return NonNullList.of(this.ingredient);
-	}
-	
-	@Override
 	public ResourceLocation getId()
 	{
 		return id;
@@ -111,12 +90,7 @@ public abstract class GristCostRecipe implements Recipe<Container>
 		return MSRecipeTypes.GRIST_COST_TYPE.get();
 	}
 	
-	public int getPriority()
-	{
-		if(priority == null)
-			return priorityFromIngredient(ingredient);
-		else return priority;
-	}
+	public abstract int getPriority();
 	
 	@Nullable
 	public abstract GristSet getGristCost(ItemStack input, @Nullable GristType wildcardType, boolean shouldRoundDown, @Nullable Level level);
@@ -130,31 +104,14 @@ public abstract class GristCostRecipe implements Recipe<Container>
 	 * Adds grist cost providers for all items which this recipe might potentially provide a grist cost for,
 	 * which then get used during grist cost generation.
 	 */
-	public void addCostProvider(BiConsumer<Item, GeneratedCostProvider> consumer)
-	{
-		GeneratedCostProvider provider = new DefaultProvider();
-		for(ItemStack stack : ingredient.getItems())
-			consumer.accept(stack.getItem(), provider);
-	}
-	
-	private class DefaultProvider implements GeneratedCostProvider
-	{
-		@Nullable
-		@Override
-		public GristCostResult generate(Item item, @Nullable GristCostResult lastCost, GenerationContext context)
-		{
-			if(lastCost == null && ingredient.test(new ItemStack(item)))
-				return new GristCostResult(getGristCost(new ItemStack(item), GristTypes.BUILD.get(), false, null));
-			else return null;
-		}
-	}
+	public abstract void addCostProvider(BiConsumer<Item, GeneratedCostProvider> consumer);
 	
 	public List<JeiGristCost> getJeiCosts(Level level)
 	{
 		return Collections.emptyList();
 	}
 	
-	private static int priorityFromIngredient(Ingredient ingredient)
+	protected static int priorityFromIngredient(Ingredient ingredient)
 	{
 		return 100 - (ingredient.getItems().length - 1)*10;
 	}
@@ -179,39 +136,5 @@ public abstract class GristCostRecipe implements Recipe<Container>
 		}
 		
 		return mutableCost;
-	}
-	
-	//Helper class for implementing serializer classes
-	public abstract static class AbstractSerializer<T extends GristCostRecipe> implements RecipeSerializer<T>
-	{
-		@Override
-		public T fromJson(ResourceLocation recipeId, JsonObject json)
-		{
-			Ingredient ingredient = Ingredient.fromJson(json.get("ingredient"));
-			Integer priority = json.has("priority") ? GsonHelper.getAsInt(json, "priority") : null;
-			
-			return read(recipeId, json, ingredient, priority);
-		}
-		
-		protected abstract T read(ResourceLocation recipeId, JsonObject json, Ingredient ingredient, @Nullable Integer priority);
-		
-		@Nullable
-		@Override
-		public T fromNetwork(ResourceLocation recipeId, FriendlyByteBuf buffer)
-		{
-			Ingredient ingredient = Ingredient.fromNetwork(buffer);
-			int priority = buffer.readInt();
-			
-			return read(recipeId, buffer, ingredient, priority);
-		}
-		
-		protected abstract T read(ResourceLocation recipeId, FriendlyByteBuf buffer, Ingredient ingredient, int priority);
-		
-		@Override
-		public void toNetwork(FriendlyByteBuf buffer, T recipe)
-		{
-			recipe.ingredient.toNetwork(buffer);
-			buffer.writeInt(recipe.getPriority());
-		}
 	}
 }
