+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/recipe/RecipeGeneratedGristCost.java
@@ -1,20 +1,18 @@
 package com.mraof.minestuck.alchemy.recipe.generator.recipe;
 
 import com.google.gson.JsonObject;
-import com.mraof.minestuck.api.alchemy.GristType;
-import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.alchemy.recipe.GristCostRecipe;
 import com.mraof.minestuck.alchemy.recipe.generator.GeneratedCostProvider;
+import com.mraof.minestuck.api.alchemy.GristSet;
+import com.mraof.minestuck.api.alchemy.GristType;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
 import com.mraof.minestuck.jei.JeiGristCost;
 import net.minecraft.MethodsReturnNonnullByDefault;
-import net.minecraft.core.NonNullList;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.level.Level;
 
@@ -33,7 +31,7 @@ public class RecipeGeneratedGristCost extends GristCostRecipe
 	
 	private RecipeGeneratedGristCost(ResourceLocation id, @Nullable RecipeGeneratedCostHandler handler)
 	{
-		super(id, Ingredient.EMPTY, -999999);	//Do not use Integer.MIN_VALUE as priority due to the risk of causing an underflow/overflow
+		super(id);
 		this.handler = handler;
 	}
 	
@@ -54,6 +52,12 @@ public class RecipeGeneratedGristCost extends GristCostRecipe
 		return getCost(inv.getItem(0).getItem()) != null;
 	}
 	
+	@Override
+	public int getPriority()
+	{
+		return -999999;	//Do not use Integer.MIN_VALUE as priority due to the risk of causing an underflow/overflow
+	}
+	
 	@Nullable
 	private GristSet getCost(Item item)
 	{
@@ -67,12 +71,6 @@ public class RecipeGeneratedGristCost extends GristCostRecipe
 	{
 		if(handler != null)
 			handler.addAsProvider(consumer);
-	}
-	
-	@Override
-	public NonNullList<Ingredient> getIngredients()
-	{
-		return NonNullList.create();
 	}
 	
 	@Override
